Adjust lifecycle hook test to check mixin override
The lifecycle hook test is adjusted to check the hook injection via
mixins is not touched by the hook override.

Furthermore the way the udpate and updated hooks are called is using the
vm.$forceUpdate method now instead of wrapper.setData, as it
consistently calls the hooks through all versions.

An callback through vm.$nextTick was added to let the update hooks be
called, which makes the test an async one using mocha's done method.