Fix event capture by installing captor via mixin
The event capture was not working when events were emitted in methods
during the component creation, like in the lifecycle hook beforeCreate.
This fix installs the captor using a mixin before the beforeCreate event
gets emitted.

The captor is available to the wrapper, when the wrapper accesses
the vue component.

Two tests were written, in emitted.spec, and emittedByOrder,
respectively.