fix: event capturing in beforeCreate lifecycle hook (#243)
* Fix event capture by installing captor via mixin

The event capture was not working when events were emitted in methods
during the component creation, like in the lifecycle hook beforeCreate.
This fix installs the captor using a mixin before the beforeCreate event
gets emitted.

The captor is available to the wrapper, when the wrapper accesses
the vue component.

Two tests were written, in emitted.spec, and emittedByOrder,
respectively.

* Adjust emitted tests to pass tested Vue versions

* Use component field for event capturing

Instead of using an external object, this.__emitted and
this.__emittedByOrder is now used to capture events. This allows the
wrapper to access the event logger of the component.

The installed event logger is never removed and therefore will exist
even after the wrapper was created.