Update lockfile
`yarn install` was failing with newer node versions (>=10) because the
version of fsevents was too old (no precompiled binaries), and node-gyp
would fail to build it.

`yarn upgrade` mitigates this issue, but a change in `prettier` 1.18
(https://prettier.io/blog/2019/06/06/1.18.0.html) required `yarn format`
to also be run.

A missing dependency (`rollup-plugin-json`) was added to
`@vue/test-utils`, as it is imported in its build script.