Remove build step from compatibility tests
Each compatibility test installs a legacy version of `vue`, and builds
each package against this version before running the test suites. This
is assumedly in error, as the only build that gets published is the one
built against the dependencies in the lockfile.

To properly test compatibility for consumers using older versions of
`vue` to satisfy each package's peer dependency, we need to test against
the build that would actually get published. The fix is to run `yarn
build:test` only once, before any compatibility tests are run.